requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
simple-salesforce==1.12.6
anthropic==0.39.0
openai==1.51.0
python-dotenv==1.0.0
pydantic==2.9.0
tiktoken==0.7.0
redis==5.0.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
